Add functions to get the elements by index of value
Move readers to traits to increase code readability
